+++ library/modules/converters/fontconverter/include/tiki/fontconverter/fontconverter.hpp
@@ -12,7 +12,7 @@ namespace tiki
 
 	protected:
 
-		virtual size_t			getConverterRevision() const { return 6u; }
+		virtual uint16			getConverterRevision() const { return 6u; }
 
 		virtual crc32			getInputType() const;
 		virtual crc32			getOutputType() const;
