+++ library/modules/converters/textureconverter/include/tiki/textureconverter/textureconverter.hpp
@@ -12,7 +12,7 @@ namespace tiki
 
 	protected:
 
-		virtual size_t			getConverterRevision() const { return 11u; }
+		virtual uint16			getConverterRevision() const { return 11u; }
 
 		virtual crc32			getInputType() const;
 		virtual crc32			getOutputType() const;
