+++ library/modules/runtime/graphics/source/win_d3d11/graphicscontext_d3d11.cpp
@@ -201,7 +201,7 @@ namespace tiki
 
 		if ( m_pDepthStencilState != pDepthStencilState )
 		{
-			m_platformData.pContext->OMSetDepthStencilState( pDepthStencilState->m_platformData.pDepthStencilState, pDepthStencilState->m_platformData.stencilRef );
+			m_platformData.pContext->OMSetDepthStencilState( pDepthStencilState->m_platformData.pDepthStencilState, UINT( pDepthStencilState->m_platformData.stencilRef ) );
 			m_pDepthStencilState = pDepthStencilState;
 		}		
 	}
